pyproject.toml
[project]
name = "projeto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.119.1",
    "uvicorn==0.38.0",
    "langchain==0.3.27",
    "langchain-community==0.3.30",
    "langchain-google-genai==2.1.12",
    "openai-whisper==20250625",
    "torch==2.8.0",
    "numpy==2.2.6",
    "python-dotenv==1.1.1",
]
